auth_client_plugin.h: silenced sign comparison warning in AUTH_CLIENT_IS_EXTENDED()
darcs-hash:20070412141540-55b16-12ad15d4ad95839cd529a1b14b706f09a9e89abd.gz

+++ libsofia-sip-ua/iptsec/sofia-sip/auth_client_plugin.h
@@ -98,7 +98,7 @@ struct auth_client_plugin
 /** Check if authentication client has been extended. @NEW_1_12_6 */
 #define AUTH_CLIENT_IS_EXTENDED(ca)					\
   ((ca)->ca_auc->auc_plugin_size >					\
-   offsetof(auth_client_plugin_t, auc_clear)				\
+   (int)offsetof(auth_client_plugin_t, auc_clear)			\
    && (ca)->ca_auc->auc_clear != NULL)
 
 SOFIA_END_DECLS
